pyproject.toml
[project]
name = "poker-settlements-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot==21.11.1",
]
